@@ -5,7 +5,7 @@ homepage: https://github.com/riczhao/xpath
 author: Richard Zhao <linuxzsc@gmail.com>
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 #dependencies:
 #  path: ^1.4.1
