@@ -1,6 +1,6 @@
 class ETree {
-  String _html;
-  Element rootElement;
+  late String _html;
+  late Element rootElement;
 
   static ETree fromString(html) {
     var tree = ETree();
@@ -13,19 +13,19 @@ class ETree {
     return tree;
   }
 
-  List<Element> xpath(String xp) => rootElement.xpath(xp);
+  List<Element>? xpath(String xp) => rootElement.xpath(xp);
 }
 
 enum ElementType { Root, Tag, Declare, Comment, Text }
 
 class Element {
-  ElementType type;
-  String name;
+  late ElementType type;
+  late String? name;
   Map<String, dynamic> attributes = {};
-  Element parent;
+  late Element parent;
   List<Element> children = [];
-  int start, end;
-  List<Element> xpath(String xp) {
+  late int start, end;
+  List<Element>? xpath(String xp) {
     var x = XPath(this, xp);
     return x.parse();
   }
@@ -41,10 +41,10 @@ var regText = RegExp(r'^[^<]*');
 var regString = RegExp(r''''.*?'|".*?"''');
 
 class BuildTree {
-  String _html;
-  int _current;
-  int _max_idx;
-  Element _parent;
+  late String _html;
+  late int _current;
+  late int _max_idx;
+  late Element _parent;
   int _depth = 0;
 
   Element rootElement = Element()..type = ElementType.Root;
@@ -161,13 +161,13 @@ class BuildTree {
 
     e.name = regName.stringMatch(s);
     if (e.name == null) return false;
-    s = s.substring(e.name.length);
+    s = s.substring(e.name!.length);
 
     // parse attributes
     while (true) {
       m = regFormatting.firstMatch(s);
       if (m != null) s = s.substring(m.end);
-      String k = regName.stringMatch(s);
+      String? k = regName.stringMatch(s);
       if (k == null) break;
       s = s.substring(k.length);
 
@@ -179,7 +179,7 @@ class BuildTree {
       //TODO: add number values
       m = regFormatting.firstMatch(s);
       if (m != null) s = s.substring(m.end);
-      String v = regString.stringMatch(s);
+      String? v = regString.stringMatch(s);
       if (v == null) break;
       s = s.substring(v.length);
       v = v.substring(1, v.length - 1);
@@ -220,7 +220,7 @@ class ScopeItem {
 
 class XPath {
   static Iterable<ScopeItem> elementAllChildren(
-      Match m, Element root, String name, List<int> i) sync* {
+      Match m, Element root, String? name, List<int> i) sync* {
     for (var e in root.children) {
       if (name != null && e.name != name) continue;
       yield ScopeItem(e, i[0]++);
@@ -251,7 +251,7 @@ class XPath {
     // /xxx
     RegExp(r'^\/([a-zA-Z0-9-_]+)'): (Match m, Element root) sync* {
       int i = 1; // xpath idx starts with 1
-      String name = m.group(1);
+      String? name = m.group(1);
       for (var e in root.children) {
         if (name != e.name) continue;
         yield ScopeItem(e, i++);
@@ -287,7 +287,7 @@ class XPath {
     // [2]
     RegExp(r'^\[([0-9]+)\]'): (Match m, Element e, int i) {
       if (m.groupCount < 1) return _MISS;
-      int idx = int.parse(m.group(1));
+      int idx = int.parse(m.group(1)!);
       if (idx == i)
         return _HIT;
       else
@@ -295,12 +295,12 @@ class XPath {
     },
     // [position()>2]
     RegExp(r'^\[position\(\)>([0-9]+)\]'): (Match m, Element e, int i) {
-      if (i > int.parse(m.group(1))) return _SELECT;
+      if (i > int.parse(m.group(1)!)) return _SELECT;
       return _MISS;
     },
     // [position()<2]
     RegExp(r'^\[position\(\)<([0-9]+)\]'): (Match m, Element e, int i) {
-      if (i < int.parse(m.group(1))) return _SELECT;
+      if (i < int.parse(m.group(1)!)) return _SELECT;
       return _MISS;
     },
   };
@@ -310,7 +310,7 @@ class XPath {
 
   XPath(this._element, this._xpath);
 
-  List<Element> parse() {
+  List<Element>? parse() {
     if (_xpath[0] != '/' || _xpath[_xpath.length - 1] == '/') return null;
 
     var selected = Set<Element>();
@@ -318,8 +318,8 @@ class XPath {
     String xpath = _xpath;
 
     while (xpath != '') {
-      Match scope_match = null;
-      RegExp scope_k = null;
+      Match? scope_match = null;
+      RegExp? scope_k = null;
       for (var k in scope.keys) {
         scope_match = k.firstMatch(xpath);
         if (scope_match != null) {
@@ -355,11 +355,11 @@ class XPath {
 
       var cur_selected = Set<Element>();
       for (var root in selected) {
-        for (var si in scope[scope_k](scope_match, root)) {
+        for (var si in scope[scope_k]!(scope_match, root)) {
           bool all_pass = true;
           bool hit = false;
           for (int i = 0; i < sel_match.length; i++) {
-            int r = selectors[sel_k[i]](sel_match[i], si.element, si.idx);
+            int r = selectors[sel_k[i]]!(sel_match[i], si.element, si.idx);
             if (r == _MISS) {
               all_pass = false;
               break;
